TradeHistory 기능 & CurrencyAsset 기능 추가

+++ backend/src/main/java/com/jkb/cmi/repository/CurrencyRepository.java
@@ -3,5 +3,5 @@ package com.jkb.cmi.repository;
 import com.jkb.cmi.entity.Currency;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurrencyRepository extends JpaRepository<Currency, Integer> {
+public interface CurrencyRepository extends JpaRepository<Currency, Long> {
 }
